프로젝트 변경, 모듈 새로 생성
