+++ .jpb/jpb-settings.xml
@@ -3,9 +3,6 @@
   <component name="PersistenceUnitSettings">
     <persistence-units>
       <persistence-unit name="Default">
-        <packages>
-          <package value="com.devlog.core" />
-        </packages>
       </persistence-unit>
     </persistence-units>
   </component>
