--- api/src/test/java/com/devlog/api/ApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.devlog.api;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
