--- core/build.gradle
@@ -1,88 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.2.2'
-    id 'io.spring.dependency-management' version '1.1.4'
-}
-
-group = 'com.devlog'
-version = '0.0.1-SNAPSHOT'
-
-java {
-    sourceCompatibility = '17'
-}
-
-configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    // security
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    // web
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    // JPA
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    // lombok
-    compileOnly 'org.projectlombok:lombok'
-    // devTools
-    developmentOnly 'org.springframework.boot:spring-boot-devtools'
-    //MySql
-    compileOnly 'mysql:mysql-connector-java'
-    compileOnly "com.querydsl:querydsl-jpa:5.0.0:jakarta"
-    compileOnly "com.querydsl:querydsl-apt:5.0.0:jakarta"
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.security:spring-security-test'
-    compileOnly "javax.validation:validation-api:2.0.1.Final"
-    runtimeOnly 'com.h2database:h2'
-
-    // Spring boot 3.x이상에서 QueryDsl 패키지를 정의하는 방법
-    // 필수
-    implementation "com.querydsl:querydsl-jpa:5.0.0:jakarta"
-    implementation "com.querydsl:querydsl-core:5.0.0"
-
-    // QueryDsl 쿼리 타입 생성 (QClass 생성 시 @Entity 탐색)
-    annotationProcessor "com.querydsl:querydsl-apt:5.0.0:jakarta"
-
-    // java.lang.NoClassDefFoundError:javax/persistence/Entity 에러 방지
-    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
-    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
-
-}
-
-
-// QueryDsl 디렉토리 경로
-def querydslDir = "$buildDir/generated/querydsl"
-
-// 경로 추가 >> QueryDsl 소스 코드 컴파일 시 빌드
-sourceSets {
-    main.java.srcDirs += [ querydslDir ]
-}
-
-// 컴파일 설정(AnnotationProcessor가 생성하는 소스코드를 해당 경로로 설정)
-tasks.withType(JavaCompile) {
-    options.annotationProcessorGeneratedSourcesDirectory = file(querydslDir)
-}
-
-// clean실행 시 마지막 작업으로 디렉토리(QClass) 삭제 >> 충돌 방지
-clean.doLast {
-    file(querydslDir).deleteDir()
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
-tasks.named('bootJar'){ // bootJar 세팅을 켜기
-    enabled = true
-}
-
-tasks.named('jar'){ // jar 설정은 끄기
-    enabled = false
-}
