--- core/src/main/java/com/devlog/core/CoreApplication.java
@@ -1,13 +1,0 @@
-package com.devlog.core;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CoreApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CoreApplication.class, args);
-    }
-
-}
