--- core/src/main/java/com/devlog/core/common/constants/CommonConstants.java
@@ -1,4 +1,0 @@
-package com.devlog.core.common.constants;
-
-public class CommonConstants {
-}
