--- core/src/main/java/com/devlog/core/common/service/entity/file/FileEntity.java
@@ -1,34 +1,0 @@
-package com.devlog.core.common.service.entity.file;
-
-import com.devlog.core.common.service.entity.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "file_mst")
-public class FileEntity extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "FILE_NO", nullable = false)
-    private Long file_no;
-
-    @Column(name = "FILE_PATH")
-    private String filePath;
-
-    @Column(name = "FILE_NAME")
-    private String fileName;
-
-    @Column(name = "FILE_EXT")
-    private String fileExt;
-
-    @Column(name = "CTNT_NO")
-    private String ctntNo;
-
-
-
-}
