--- core/src/main/java/com/devlog/core/common/service/entity/user/UserEntity.java
@@ -1,32 +1,0 @@
-package com.devlog.core.common.service.entity.user;
-
-import com.devlog.core.common.service.entity.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "user_mst")
-public class UserEntity extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "USER_NO", nullable = false)
-    private Long userNo;
-
-    @Column(name = "USERNAME")
-    private String username;
-
-    @Column(name = "EMAIL")
-    private String email;
-
-    @Column(name = "PASSWORD")
-    private String password;
-
-    @Column(name = "GRANT")
-    private String grant;
-
-}
