--- core/src/main/java/com/devlog/core/common/service/repository/content/ContentRepository.java
@@ -1,9 +1,0 @@
-package com.devlog.core.common.service.repository.content;
-
-import com.devlog.core.common.service.entity.content.ContentEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ContentRepository extends CrudRepository<ContentEntity, Long> {
-}
