--- core/src/test/java/com/devlog/core/CoreApplicationTests.java
@@ -1,13 +1,0 @@
-package com.devlog.core;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CoreApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
