+++ ztlog-core/src/main/java/com/devlog/core/common/constants/CommonConstants.java
@@ -1,0 +1,10 @@
+package com.devlog.core.common.constants;
+
+public class CommonConstants {
+    public static final String TITLE = "ztlog";
+
+    public static final String USER_NOT_FOUND = "사용자 정보가 없습니다.";
+
+    public static final String MEMBER_ERROR_PASSWORD = "자격 증명에 실패하였습니다.";
+
+}
