+++ ztlog-core/src/main/java/com/devlog/core/config/JPAConfig.java
@@ -1,0 +1,11 @@
+package com.devlog.core.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EnableJpaAuditing
+@EnableJpaRepositories(basePackages = "com.devlog.core.domain.repository.*")
+public class JPAConfig {
+}
