+++ ztlog-core/src/main/java/com/devlog/core/config/exception/DataNotFoundException.java
