+++ ztlog-core/src/main/java/com/devlog/core/domain/entity/BaseTimeEntity.java
@@ -1,0 +1,27 @@
+package com.devlog.core.domain.entity;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public class BaseTimeEntity {
+
+    @CreatedDate
+    @Column(name = "INP_DTTM", nullable = false, updatable = false)
+    private LocalDateTime inpDttm;
+
+    @LastModifiedDate
+    @Column(name = "UPD_DTTM")
+    private LocalDateTime updDttm;
+
+}
