+++ ztlog-core/src/main/java/com/devlog/core/domain/repository/content/ContentRepository.java
@@ -1,0 +1,9 @@
+package com.devlog.core.domain.repository.content;
+
+import com.devlog.core.domain.entity.content.ContentEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ContentRepository extends CrudRepository<ContentEntity, Long> {
+}
