+++ ztlog-core/src/main/java/com/devlog/core/domain/repository/file/FileRepository.java
@@ -1,0 +1,9 @@
+package com.devlog.core.domain.repository.file;
+
+import com.devlog.core.domain.entity.file.FileEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface FileRepository extends CrudRepository<FileEntity, Long> {
+}
