+++ ztlog-core/src/main/java/com/devlog/core/domain/repository/user/UserRepository.java
@@ -1,0 +1,11 @@
+package com.devlog.core.domain.repository.user;
+
+import com.devlog.core.domain.entity.user.UserEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@ResponseBody
+public interface UserRepository extends CrudRepository<UserEntity, Long> {
+
+    UserEntity findByUsername(String username);
+}
