+++ ztlog-core/src/main/resources/application.properties
