+++ ztlog-core/src/test/java/com/devlog/core/ZtlogCoreApplicationTests.java
@@ -1,0 +1,13 @@
+package com.devlog.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ZtlogCoreApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
